+++ com.unity.postprocessing/PostProcessing/Editor/Effects/AutoExposureEditor.cs
@@ -32,7 +32,17 @@ namespace UnityEditor.Rendering.PostProcessing
         public override void OnInspectorGUI()
         {
             if (!SystemInfo.supportsComputeShaders)
-                EditorGUILayout.HelpBox("Auto exposure requires compute shader support.", MessageType.Warning);
+            {
+                EditorGUILayout.HelpBox("Auto exposure requires compute shader support which is not available on this platform.", MessageType.Error);
+            }
+            else if (EditorUtilities.isTargetingWebGL)
+            {
+                EditorGUILayout.HelpBox("Auto exposure requires compute shader support (WebGPU) when running on Web.", MessageType.Warning);
+            }
+            else if (EditorUtilities.isTargetingAndroid)
+            {
+                EditorGUILayout.HelpBox("Auto exposure requires compute shader support (Vulkan) when running on Android.", MessageType.Warning);
+            }
 
             EditorUtilities.DrawHeaderLabel("Exposure");