+++ com.unity.template-hd/Packages/com.unity.template.hd/Tests/Editor/InternalTests/ExpectedSettings.cs
@@ -16,10 +16,8 @@ namespace Tests
             {
                 yield return new TestCaseData(BuildTarget.Android);
                 yield return new TestCaseData(BuildTarget.iOS);
-                yield return new TestCaseData(BuildTarget.Lumin);
                 yield return new TestCaseData(BuildTarget.PS4);
                 yield return new TestCaseData(BuildTarget.PS5);
-                yield return new TestCaseData(BuildTarget.Stadia);
                 yield return new TestCaseData(BuildTarget.StandaloneLinux64);
                 yield return new TestCaseData(BuildTarget.StandaloneOSX);
                 yield return new TestCaseData(BuildTarget.StandaloneWindows);
@@ -41,13 +39,11 @@ namespace Tests
                 case BuildTarget.StandaloneOSX:
                 case BuildTarget.Android:
                 case BuildTarget.iOS:
-                case BuildTarget.Lumin:
                 case BuildTarget.tvOS:
                 case BuildTarget.WebGL:
                     return false;
                 case BuildTarget.PS4:
                 case BuildTarget.PS5:
-                case BuildTarget.Stadia:
                 case BuildTarget.StandaloneLinux64:
                 case BuildTarget.StandaloneWindows:
                 case BuildTarget.StandaloneWindows64: